move `IncrementalSchemaLink` and others to `shared`
