+++ integration-test/nextjs/src/app/cc/ApolloWrapper.tsx
@@ -10,11 +10,11 @@ import {
 import { loadErrorMessages, loadDevMessages } from "@apollo/client/dev";
 import { setVerbosity } from "ts-invariant";
 import { delayLink } from "@/shared/delayLink";
-import { schema } from "../graphql/schema";
+import { schema } from "@integration-test/shared/schema";
 
 import { useSSROnlySecret } from "ssr-only-secrets";
 import { errorLink } from "../../shared/errorLink";
-import { IncrementalSchemaLink } from "../graphql/IncrementalSchemaLink";
+import { IncrementalSchemaLink } from "@integration-test/shared/IncrementalSchemaLink";
 
 setVerbosity("debug");
 loadDevMessages();
