+++ integration-test/nextjs/src/app/rsc/client.ts
@@ -9,8 +9,8 @@ import { setVerbosity } from "ts-invariant";
 import { delayLink } from "@/shared/delayLink";
 import { errorLink } from "@/shared/errorLink";
 
-import { schema } from "../graphql/schema";
-import { IncrementalSchemaLink } from "../graphql/IncrementalSchemaLink";
+import { schema } from "@integration-test/shared/schema";
+import { IncrementalSchemaLink } from "@integration-test/shared/IncrementalSchemaLink";
 
 setVerbosity("debug");
 loadDevMessages();
